requirements.txt
flask
python-dotenv
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
validators
Flask-Limiter
redis